Port TMV tiled/mip/16k work onto current upstream main
Replays the TMV feature branch (tmv_tiled_output, 51 commits since the
#139 fork point) on top of AcademySoftwareFoundation/openapv main, which
had diverged with a codebase-wide API rename and input-validation
hardening. The TMV line had been developed on a re-initialized git
history with no shared ancestry, so this is a port rather than a merge.

Conflict resolution highlights:
- Adopt upstream's renamed encoder/decoder API (num_c, c_sft,
  fn_blk_from_pic/fn_blk_to_pic, oapv_validate_tile_topology,
  4-arg args_set_variable_by_key_long, 5-arg oapvd_vlc_tile_header,
  3-arg oapv_bsr_read_direct, 3-arg dec_frm_prepare) throughout the
  TMV-added code.
- Keep TMV features: dynamic tile arrays (encoder + decoder), per-frame
  RC (rc_param_frm), fn_imgb_to_blk_rc sampler, selective/disk-I/O
  multi-mip decoder + tile-status enum + tile buffer structs, --tmv-mips
  encoder option, per-AU output files, 64x64 tile limits.
- Keep upstream hardening: ndata[0] bitstream-overflow flag path,
  fps/band validation, payload-bounds checks, 16-bit imgb_write range.
- Keep TMV's total_bits u64 overflow fix (16k rate control).
- Recognize .oapv (not just .apv) as a single-file output extension so
  the per-AU split path only triggers for extension-less output names.
- Add dec_frm_prepare_selective for the selective/multi-mip decode paths:
  upstream's hardened dec_frm_prepare validates the imgb against frame
  dimensions/format/capacity, which rejected the dummy imgb those paths
  pass (they decode into their own work buffers). The selective variant
  sets up ctx metadata + tile geometry without imgb validation, leaving
  upstream's dec_frm_prepare intact for the normal full-frame path.

Builds clean (gcc 13, all targets). Encode/decode roundtrip validated
(125-frame 320x240 422p10). All 16 ctest tests pass. Selective multi-mip
decode API (oapvd_decode_selective_multi_mips) validated via decoder_test
run_4k_tests set: all real cases decode correctly; the only non-OK
results are the intentional negative tests (non-existent mip -> NOT_FOUND)
and a sparse-corner layout that trips the validator's zero-column heuristic.

Original 51-commit TMV history preserved in branch pr_candidate_backup.

+++ src/oapv_rc.c
@@ -214,9 +214,12 @@ void oapve_rc_update_after_pic(oapve_ctx_t* ctx, double cost)
             num_pixel += (ctx->w * ctx->h) >> (ctx->c_sft[c][0] + ctx->c_sft[c][1]);
         }
 
-        int total_bits = 0;
+        /* total_bits must be u64: a 16k 10-bit 4:2:2 frame at high quality can
+         * exceed INT_MAX bits (~33 MB encoded), which previously wrapped negative
+         * and turned the subsequent log() into NaN, corrupting alpha/beta. */
+        u64 total_bits = 0;
         for (int i = 0; i < ctx->num_tiles; i++) {
-            total_bits += ctx->fh.tile_size[i] * 8;
+            total_bits += (u64)ctx->fh.tile_size[i] * 8;
         }
 
         double ln_bpp = log(pow(cost / (double)num_pixel, OAPV_RC_BETA));
